Fix connect tab performance
