--- src/components/PeopleSearchResults/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "PeopleSearchResults.js",
-  "version": "0.0.0",
-  "private": true,
-  "main": "./PeopleSearchResults.js"
-}
