+++ src/screens/People/People.js
@@ -20,7 +20,7 @@
 
 import * as React from 'react';
 import { connect } from 'react-redux';
-import { FlatList, Keyboard, RefreshControl, View, ScrollView } from 'react-native';
+import { FlatList, Keyboard, RefreshControl } from 'react-native';
 import Swipeout from 'react-native-swipeout';
 import debounce from 'lodash.debounce';
 import isEqual from 'lodash.isequal';
@@ -36,7 +36,11 @@ import {
   muteContactAction,
   blockContactAction,
 } from 'actions/contactsActions';
-import { fetchInviteNotificationsAction } from 'actions/invitationsActions';
+import { fetchInviteNotificationsAction,
+  sendInvitationAction,
+  acceptInvitationAction,
+  cancelInvitationAction,
+  rejectInvitationAction } from 'actions/invitationsActions';
 import { logScreenViewAction } from 'actions/analyticsActions';
 import { goToInvitationFlowAction } from 'actions/referralsActions';
 
@@ -46,17 +50,17 @@ import { Wrapper } from 'components/Layout';
 import SearchBlock from 'components/SearchBlock';
 import ListItemWithImage from 'components/ListItem/ListItemWithImage';
 import Spinner from 'components/Spinner';
-import { BaseText } from 'components/Typography';
+import { BaseText, SubHeadingMedium } from 'components/Typography';
 import NotificationCircle from 'components/NotificationCircle';
 import Button from 'components/Button';
-import PeopleSearchResults from 'components/PeopleSearchResults';
 import EmptyStateParagraph from 'components/EmptyState/EmptyStateParagraph';
 import ContainerWithHeader from 'components/Layout/ContainerWithHeader';
 import ConnectionConfirmationModal from 'screens/Contact/ConnectionConfirmationModal';
+import Overlay from 'components/SearchBlock/Overlay';
 
 // constants
 import { CONTACT, CONNECTION_REQUESTS } from 'constants/navigationConstants';
-import { TYPE_RECEIVED } from 'constants/invitationsConstants';
+import { TYPE_RECEIVED, TYPE_INVITE, TYPE_REJECTED, TYPE_SENT } from 'constants/invitationsConstants';
 import {
   DISCONNECT,
   MUTE,
@@ -67,13 +71,19 @@ import {
 
 // models/types
 import type { Dispatch, RootReducerState } from 'reducers/rootReducer';
-import type { SearchResults } from 'models/Contacts';
+import type { SearchResults, ApiUser } from 'models/Contacts';
 import type { Theme } from 'models/Theme';
 
 // utils
-import { fontSizes, spacing, fontStyles } from 'utils/variables';
+import { fontSizes, spacing, fontStyles, itemSizes } from 'utils/variables';
 import { getThemeColors, themedColors } from 'utils/themes';
 import { sortLocalContacts } from 'utils/contacts';
+
+import unionBy from 'lodash.unionby';
+import intersectionBy from 'lodash.intersectionby';
+import Separator from 'components/Separator';
+import ProfileImage from 'components/ProfileImage';
+import { createAlert } from 'utils/alerts';
 
 // partials
 import InviteBanner from './InviteBanner';
@@ -132,7 +142,40 @@ const EmptyStateWrapper = styled.View`
   justify-content: center;
 `;
 
+const LocalContacts = styled.View`
+  border-top-width: 1px;
+  border-bottom-width: 1px;
+  border-style: solid;
+  border-color: ${themedColors.border};
+`;
+
+const LocalContactsScrollView = styled.ScrollView`
+  margin-bottom: ${spacing.small}px;
+`;
+
+const LocalContactsSubHeading = styled(SubHeadingMedium)`
+  margin: 22px 16px 0;
+`;
+
+const ListSubHeading = styled(SubHeadingMedium)`
+  margin: 6px ${spacing.mediumLarge}px 8px;
+`;
+
+const LocalContactsItem = styled.TouchableOpacity`
+  align-items: center;
+  width: ${itemSizes.avatarCircleMedium + 4}px;
+  margin: 0 8px;
+`;
+
+const LocalContactsItemName = styled(BaseText)`
+  ${fontStyles.small};
+  color: ${themedColors.secondaryText};
+  padding: 0 4px;
+  margin-top: 3px;
+`;
+
 const MIN_QUERY_LENGTH = 2;
+const OVERLAY_OFFSET = 72;
 
 type Props = {
   navigation: NavigationScreenProp<*>,
@@ -151,6 +194,10 @@ type Props = {
   theme: Theme,
   referralsFeatureEnabled: boolean,
   goToInvitationFlow: () => void,
+  sendInvitation: Function,
+  acceptInvitation: Function,
+  cancelInvitation: Function,
+  rejectInvitation: Function,
 }
 
 type ConnectionStatusProps = {
@@ -163,6 +210,7 @@ type State = {
   manageContactType: string,
   manageContactId: string,
   forceHideRemoval: boolean,
+  isSearchFocused: boolean,
 }
 
 const ConnectionStatus = (props: ConnectionStatusProps) => {
@@ -187,8 +235,9 @@ const ConnectionStatus = (props: ConnectionStatusProps) => {
 class PeopleScreen extends React.Component<Props, State> {
   didBlur: NavigationEventSubscription;
   willFocus: NavigationEventSubscription;
-  scrollViewRef: ScrollView;
   flatListRef: FlatList;
+  searchBarRef: Object;
+
   forceRender = false;
   state = {
     query: '',
@@ -196,12 +245,14 @@ class PeopleScreen extends React.Component<Props, State> {
     manageContactType: '',
     manageContactId: '',
     forceHideRemoval: false,
+    isSearchFocused: false,
   };
 
   constructor(props: Props) {
     super(props);
     this.handleSearchChange = debounce(this.handleSearchChange, 500);
-    this.scrollViewRef = React.createRef();
+    this.searchBarRef = React.createRef();
+    this.flatListRef = React.createRef();
   }
 
   componentDidMount() {
@@ -326,7 +377,6 @@ class PeopleScreen extends React.Component<Props, State> {
         close={this.state.forceHideRemoval}
         buttonWidth={80}
         scroll={(shouldAllowScroll) => {
-          this.toggleScroll(this.scrollViewRef, shouldAllowScroll);
           this.toggleScroll(this.flatListRef, shouldAllowScroll);
         }}
       >
@@ -343,6 +393,30 @@ class PeopleScreen extends React.Component<Props, State> {
           noSeparator
         />
       </Swipeout>
+    );
+  };
+
+  renderSearchModeContact = ({ item: user }) => {
+    const { invitations, navigation } = this.props;
+    const invitation = invitations.find(({ id }) => id === user.id);
+    let status = TYPE_INVITE;
+    if (invitation) {
+      status = invitation.type;
+    }
+
+    return (
+      <ListItemWithImage
+        label={user.username}
+        avatarUrl={user.profileImage}
+        navigateToProfile={() => navigation.navigate(CONTACT, { contact: user })}
+        rejectInvitation={this.handleRejectInvitationPress(user)}
+        acceptInvitation={this.handleAcceptInvitationPress(user)}
+        buttonAction={status === TYPE_SENT
+          ? this.handleCancelInvitationPress(user)
+          : this.handleSendInvitationPress(user)}
+        buttonActionLabel={status === TYPE_SENT ? 'Requested' : 'Connect'}
+        secondaryButton={status === TYPE_SENT}
+      />
     );
   };
 
@@ -396,92 +470,200 @@ class PeopleScreen extends React.Component<Props, State> {
     );
   };
 
-  renderContent = (sortedLocalContacts: Object[], inSearchMode: boolean) => {
+  renderSearchBlock = () => {
     const {
-      searchResults,
-      searchResults: { apiUsers, localContacts },
       isSearching,
-      navigation,
-      invitations,
-      chats,
-      theme,
+    } = this.props;
+    return (
+      <SearchBlock
+        headerProps={{ title: 'people' }}
+        searchInputPlaceholder="Search or add people"
+        onSearchChange={(q) => this.handleSearchChange(q)}
+        onSearchFocus={() => {
+          this.setState({ isSearchFocused: true });
+          this.flatListRef.current.scrollToOffset({ offset: 0 });
+        }}
+        onSearchBlur={() => this.setState({ isSearchFocused: false })}
+        itemSearchState={isSearching}
+        wrapperStyle={{ paddingHorizontal: spacing.layoutSides, paddingVertical: spacing.mediumLarge }}
+        hideOverlay
+        ref={this.searchBarRef}
+      />
+    );
+  }
+
+  handleSendInvitationPress = (user: ApiUser) => () => {
+    Keyboard.dismiss();
+    this.props.sendInvitation(user);
+  };
+
+  handleAcceptInvitationPress = (user: ApiUser) => () => {
+    const { acceptInvitation, invitations } = this.props;
+    const invitation = invitations.find(({ id }) => id === user.id);
+    Keyboard.dismiss();
+    acceptInvitation(invitation);
+  };
+
+  handleCancelInvitationPress = (user: ApiUser) => () => {
+    const { cancelInvitation, invitations } = this.props;
+    const invitation = invitations.find(({ id }) => id === user.id);
+    Keyboard.dismiss();
+    cancelInvitation(invitation);
+  };
+
+  handleRejectInvitationPress = (user: ApiUser) => () => {
+    const { rejectInvitation, invitations } = this.props;
+    const invitation = invitations.find(({ id }) => id === user.id);
+    Keyboard.dismiss();
+    if (invitation && Object.keys(invitation).length > 0) {
+      createAlert(TYPE_REJECTED, invitation, () => rejectInvitation(invitation));
+    }
+  };
+
+  renderLocalContactsList = () => {
+    const {
+      searchResults: { apiUsers, localContacts: resultsLocalContacts },
+      localContacts,
     } = this.props;
 
-    const usersFound = (apiUsers.length + localContacts.length) > 0;
+    const updatedLocalContact = intersectionBy(localContacts, apiUsers, 'id');
+    const filteredLocalContacts = unionBy(resultsLocalContacts, updatedLocalContact, 'id');
+
+    if (filteredLocalContacts.length) {
+      return (
+        <LocalContacts>
+          <LocalContactsSubHeading>MY CONTACTS</LocalContactsSubHeading>
+          <LocalContactsScrollView
+            keyboardShouldPersistTaps="always"
+            horizontal
+            contentContainerStyle={{ paddingHorizontal: spacing.large / 2, paddingVertical: spacing.medium }}
+          >
+            {this.renderLocalContacts(filteredLocalContacts)}
+          </LocalContactsScrollView>
+        </LocalContacts>
+      );
+    }
+    return null;
+  };
+
+  renderLocalContacts = (filteredLocalContacts) => {
+    const { navigation } = this.props;
+    return filteredLocalContacts
+      .map(contact => (
+        <LocalContactsItem
+          key={contact.username}
+          onPress={() => navigation.navigate(CONTACT, { contact })}
+        >
+          <ProfileImage
+            uri={contact.profileImage}
+            userName={contact.username}
+            diameter={itemSizes.avatarCircleMedium}
+            textStyle={{ fontSize: fontSizes.big }}
+            noShadow
+            borderWidth={0}
+          />
+          <LocalContactsItemName numberOfLines={1}>{contact.username}</LocalContactsItemName>
+        </LocalContactsItem>
+      ));
+  };
+
+  renderConnectionRequests = () => {
+    const {
+      invitations,
+      theme,
+    } = this.props;
     const pendingConnectionRequests = invitations.filter(({ type }) => type === TYPE_RECEIVED).length;
     const colors = getThemeColors(theme);
 
-    return (
-      <React.Fragment>
-        <SearchBlock
-          headerProps={{ title: 'people' }}
-          searchInputPlaceholder="Search or add people"
-          onSearchChange={(q) => this.handleSearchChange(q)}
-          itemSearchState={isSearching}
-          wrapperStyle={{ paddingHorizontal: spacing.layoutSides, paddingVertical: spacing.mediumLarge }}
-        />
-        {!inSearchMode && !!pendingConnectionRequests &&
-        <ConnectionRequestBanner
-          onPress={this.handleConnectionsRequestBannerPress}
-          underlayColor={colors.secondaryAccent}
-        >
-          <React.Fragment>
-            <ConnectionRequestBannerText>
-              Connection requests
-            </ConnectionRequestBannerText>
-            <ConnectionRequestNotificationCircle>
-              {pendingConnectionRequests}
-            </ConnectionRequestNotificationCircle>
-            <ConnectionRequestBannerIcon name="chevron-right" />
-          </React.Fragment>
-        </ConnectionRequestBanner>
-        }
-        <InnerWrapper>
-          {inSearchMode && usersFound &&
-          <PeopleSearchResults
-            searchResults={searchResults}
-            navigation={navigation}
-            invitations={invitations}
-            localContacts={sortedLocalContacts}
-          />
-          }
-          {!inSearchMode && !!sortedLocalContacts.length &&
-          <FlatList
-            ref={(ref) => { this.flatListRef = ref; }}
-            data={sortedLocalContacts}
-            extraData={chats}
-            keyExtractor={(item) => item.id}
-            renderItem={this.renderContact}
-            initialNumToRender={8}
-            onScroll={() => Keyboard.dismiss()}
-            contentContainerStyle={{
-              paddingVertical: spacing.rhythm,
-              paddingTop: 0,
-            }}
-          />
-          }
-          {(!inSearchMode || !apiUsers.length) &&
-          <View
-            style={{ flex: 1 }}
-          >
-            {isSearching && <Wrapper center style={{ flex: 1 }}><Spinner /></Wrapper>}
+    if (!pendingConnectionRequests) return null;
 
-            {inSearchMode && !isSearching && !usersFound && this.renderEmptyState({
-              inviteTitle: 'Pillar is social',
-              esTitle: 'Nobody found',
-              esBody: 'Make sure you entered the name correctly',
-            })}
-            {!inSearchMode && !sortedLocalContacts.length && this.renderEmptyState({
-              inviteTitle: 'Invite friends',
-              esTitle: 'Start making friends',
-              esBody: 'Build your connection list by searching for someone',
-            })}
-          </View>
-          }
-        </InnerWrapper>
-      </React.Fragment>
+    return (
+      <ConnectionRequestBanner
+        onPress={this.handleConnectionsRequestBannerPress}
+        underlayColor={colors.secondaryAccent}
+      >
+        <React.Fragment>
+          <ConnectionRequestBannerText>
+            Connection requests
+          </ConnectionRequestBannerText>
+          <ConnectionRequestNotificationCircle>
+            {pendingConnectionRequests}
+          </ConnectionRequestNotificationCircle>
+          <ConnectionRequestBannerIcon name="chevron-right" />
+        </React.Fragment>
+      </ConnectionRequestBanner>
     );
-  };
+  }
+
+  renderContent = (onScroll) => {
+    const {
+      query,
+    } = this.state;
+    const {
+      localContacts,
+      chats,
+      fetchInviteNotifications,
+      searchResults: { apiUsers },
+      isSearching,
+    } = this.props;
+    const inSearchMode = query.length >= MIN_QUERY_LENGTH;
+
+    const sortedLocalContacts = sortLocalContacts(localContacts, chats);
+    const localContactsIds = sortedLocalContacts.map(({ id }) => id);
+    const filteredApiUsers = apiUsers.filter((user) => !localContactsIds.includes(user.id));
+
+    return (
+      <FlatList
+        ref={this.flatListRef}
+        data={inSearchMode ? filteredApiUsers : sortedLocalContacts}
+        renderItem={inSearchMode ? this.renderSearchModeContact : this.renderContact}
+        keyExtractor={({ username }) => username}
+        onScroll={onScroll}
+        scrollEventThrottle={16}
+        keyboardShouldPersistTaps="never"
+        contentContainerStyle={{
+          flexGrow: 1,
+        }}
+        extraData={chats}
+        refreshControl={
+          <RefreshControl
+            refreshing={false}
+            onRefresh={() => { fetchInviteNotifications(); }}
+          />
+          }
+        ListEmptyComponent={
+          <>
+            {inSearchMode && !isSearching && this.renderEmptyState({
+                inviteTitle: 'Pillar is social',
+                esTitle: 'Nobody found',
+                esBody: 'Make sure you entered the name correctly',
+              })}
+            {!inSearchMode && !isSearching && this.renderEmptyState({
+                inviteTitle: 'Invite friends',
+                esTitle: 'Start making friends',
+                esBody: 'Build your connection list by searching for someone',
+              })}
+            {isSearching && <Wrapper center style={{ flex: 1 }}><Spinner /></Wrapper>}
+          </>
+          }
+        ListHeaderComponent={
+          <>
+            {this.renderSearchBlock()}
+            {!inSearchMode && this.renderConnectionRequests()}
+            {inSearchMode && this.renderLocalContactsList()}
+            {inSearchMode && !!filteredApiUsers.length && <ListSubHeading>ALL USERS</ListSubHeading>}
+          </>
+          }
+        ItemSeparatorComponent={() => inSearchMode ? <Separator spaceOnLeft={82} /> : null}
+      />
+    );
+  }
+
+  handleOverlayClick = () => {
+    this.handleSearchChange('');
+    this.setState({ isSearchFocused: false });
+    this.searchBarRef.current.handleSearchBlur();
+  }
 
   render() {
     const {
@@ -489,11 +671,11 @@ class PeopleScreen extends React.Component<Props, State> {
       showConfirmationModal,
       manageContactType,
       manageContactId,
+      isSearchFocused,
     } = this.state;
     const {
       localContacts,
       chats,
-      fetchInviteNotifications,
     } = this.props;
     const inSearchMode = query.length >= MIN_QUERY_LENGTH;
 
@@ -507,25 +689,8 @@ class PeopleScreen extends React.Component<Props, State> {
         tab
       >
         {onScroll => (
-          <ScrollView
-            keyboardShouldPersistTaps="always"
-            contentContainerStyle={{ flexGrow: 1 }}
-            ref={(ref) => { this.scrollViewRef = ref; }}
-            onScroll={(ev) => {
-              if (inSearchMode) {
-                Keyboard.dismiss();
-              }
-              onScroll(ev);
-            }}
-            scrollEventThrottle={16}
-            refreshControl={
-              <RefreshControl
-                refreshing={false}
-                onRefresh={() => { fetchInviteNotifications(); }}
-              />
-          }
-          >
-            {this.renderContent(sortedLocalContacts, inSearchMode)}
+          <InnerWrapper>
+            {this.renderContent(onScroll)}
             {contact && <ConnectionConfirmationModal
               showConfirmationModal={showConfirmationModal}
               manageContactType={manageContactType}
@@ -538,7 +703,12 @@ class PeopleScreen extends React.Component<Props, State> {
               });
             }}
             />}
-          </ScrollView>
+            <Overlay
+              active={isSearchFocused && !inSearchMode}
+              topOffset={OVERLAY_OFFSET}
+              handleClick={this.handleOverlayClick}
+            />
+          </InnerWrapper>
         )}
       </ContainerWithHeader>
     );
@@ -574,6 +744,10 @@ const mapDispatchToProps = (dispatch: Dispatch): $Shape<Props> => ({
   blockContact: (contactId: string, block: boolean) => dispatch(blockContactAction(contactId, block)),
   logScreenView: (view: string, screen: string) => dispatch(logScreenViewAction(view, screen)),
   goToInvitationFlow: () => dispatch(goToInvitationFlowAction()),
+  sendInvitation: (user) => dispatch(sendInvitationAction(user)),
+  acceptInvitation: (user) => dispatch(acceptInvitationAction(user)),
+  cancelInvitation: (user) => dispatch(cancelInvitationAction(user)),
+  rejectInvitation: (user) => dispatch(rejectInvitationAction(user)),
 });
 
 export default withTheme(connect(mapStateToProps, mapDispatchToProps)(PeopleScreen));
